+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # 2.0.0
 
 - Smaller codebase
-- Added in new middlewares (`sts`, `rds-signer`, `http-server-timing`)
-- Added new profiler mode for `core` to allow easier finding of bottlenecks with middlewares
+- Added in new middlewares (`http-server-timing`, `rds-signer`, `sts`)
+- Added new profiler mode for `core` to allow easier bottlenecks detection with middlewares and handler
 
 ## Breaking Changes
 - Updated `aws-sdk` to v3
 - Updated all packages to be ES6 modules
-- All middleware now use `async/await` and have deprecated `next()` and `callback()`
+- All middlewares now use `async/await` and have deprecated `next()` and `callback()`
 - Error handling has been a common pain point for some, it's been reworked
 - `validator` refactored to support `draft-2019-09` using the latest version of `ajv`. Full `i18n` is now enabled by default (MAYBE)
 - Middlewares that reach out to 3rd party API have been completely refactored to have unifying options. Applies to: